refactor: 여러 페이지에서 사용되는 시간 유틸 함수 분리 #57
